requirements.txt
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pyyaml==6.0.2
python-dotenv==1.0.1
xgboost==2.0.3
scikit-learn==1.5.1
shap==0.45.1
lightgbm==4.3.0
# dbt core + postgres adapter
Jinja2==3.1.4
dbt-core==1.7.14
dbt-postgres==1.7.14
# Airflow (pin to reduce bloat if using docker-compose services instead)
apache-airflow==2.9.3